Minor refactoring for screens and navigation
