+++ chaewon_login/db/db_manager.py
@@ -3,7 +3,7 @@ from enum import Enum
 
 from chaewon_login.db.mongo import get_collection
 from chaewon_login.db.sqlite import connect_to_sqlite, find_user_sqlite, insert_user_sqlite
-from chaewon_login.ui.loading_screen import show_loading_screen
+from chaewon_login.ui.screens.loading_screen import show_loading_screen
 from typing import Callable
 
 
@@ -73,7 +73,7 @@ def init_database(page=None, callback: Callable=None):
         initialized = True
 
         if page:
-            from chaewon_login.ui.login_ui import main_login_ui
+            from chaewon_login.ui.screens.login_ui import main_login_ui
             page.controls.clear()
             page.overlay.clear()
             page.update()
