+++ chaewon_login/main.py
@@ -1,15 +1,15 @@
 import flet as ft
 
 from chaewon_login.ui.styles import apply_default_page_config
-from chaewon_login.ui.route_handling import (
-    ROUTE_HANDLERS,
-    handle_not_found,
-    match_dynamic_route
+from chaewon_login.routing.route_handling import (
+    ROUTE_HANDLERS, handle_not_found, match_dynamic_route
 )
-from chaewon_login.ui.route_data import PageRoute
+from chaewon_login.routing.route_data import PageRoute
 from chaewon_login.ui.transitions import fade_in
 from chaewon_login.auth.user import is_authenticated
 
+
+LOGIN_PAGE = PageRoute.LOGIN.value
 
 def main(page: ft.Page):
     page.title = "Chaewon's Meet and Greet"
@@ -21,14 +21,14 @@ def main(page: ft.Page):
         route = ROUTE_HANDLERS.get(page.route)
         if route:
             if route.auth_required and not is_authenticated(page):
-                page.go(PageRoute.LOGIN.value)
+                page.go(LOGIN_PAGE)
                 return
             route.handler(page, e)
         else:
             dynamic, params = match_dynamic_route(page.route)
             if dynamic:
                 if dynamic["auth_required"] and not is_authenticated(page):
-                    page.go(PageRoute.LOGIN.value)
+                    page.go(LOGIN_PAGE)
                     return
                 dynamic["handler"](page, e, **params)
             else:
@@ -38,6 +38,6 @@ def main(page: ft.Page):
 
 
     page.on_route_change = route_change
-    page.go(page.route or PageRoute.LOADING.value)
+    page.go(page.route or LOGIN_PAGE)
 
 ft.app(target=main, assets_dir="chaewon_login/assets")
